refactored settings up a level. Fixed bug in BTree.
